+++ contracts/mocks/token/ERC1155721/ERC1155721InventoryPausableMock.sol
@@ -12,10 +12,11 @@ import {Pausable} from "@animoca/ethereum-contracts-core-1.1.2/contracts/lifecyc
  * @dev The minting functions are usable while paused as it can be useful for contract maintenance such as contract migration.
  */
 contract ERC1155721InventoryPausableMock is Pausable, ERC1155721InventoryBurnableMock {
-    constructor(IForwarderRegistry forwarderRegistry, address universalForwarder)
-        ERC1155721InventoryBurnableMock(forwarderRegistry, universalForwarder)
-        Pausable(false)
-    {}
+    constructor(
+        IForwarderRegistry forwarderRegistry,
+        address universalForwarder,
+        uint256 collectionMaskLength
+    ) ERC1155721InventoryBurnableMock(forwarderRegistry, universalForwarder, collectionMaskLength) Pausable(false) {}
 
     //================================================== Pausable (admin) ===================================================//