+++ contracts/token/ERC1155/ERC1155InventoryBase.sol
@@ -19,6 +19,8 @@ import {IERC1155TokenReceiver} from "./../ERC1155/IERC1155TokenReceiver.sol";
 abstract contract ERC1155InventoryBase is ManagedIdentity, IERC165, IERC1155Inventory, IERC1155MetadataURI, IERC1155InventoryTotalSupply {
     using ERC1155InventoryIdentifiersLib for uint256;
 
+    uint256 internal immutable _collectionMaskLength;
+
     // bytes4(keccak256("onERC1155Received(address,address,uint256,uint256,bytes)"))
     bytes4 internal constant _ERC1155_RECEIVED = 0xf23a6e61;
 
@@ -43,6 +45,11 @@ abstract contract ERC1155InventoryBase is ManagedIdentity, IERC165, IERC1155Inve
     /* collection ID => creator */
     mapping(uint256 => address) internal _creators;
 
+    constructor(uint256 collectionMaskLength) {
+        require(collectionMaskLength != 0 && collectionMaskLength < 256, "Inventory: wrong mask length");
+        _collectionMaskLength = collectionMaskLength;
+    }
+
     //======================================================= ERC165 ========================================================//
 
     /// @inheritdoc IERC165
@@ -61,7 +68,7 @@ abstract contract ERC1155InventoryBase is ManagedIdentity, IERC165, IERC1155Inve
     function balanceOf(address owner, uint256 id) public view virtual override returns (uint256) {
         require(owner != address(0), "Inventory: zero address");
 
-        if (id.isNonFungibleToken()) {
+        if (id.isNonFungibleToken(_collectionMaskLength)) {
             return address(uint160(_owners[id])) == owner ? 1 : 0;
         }
 
@@ -102,9 +109,9 @@ abstract contract ERC1155InventoryBase is ManagedIdentity, IERC165, IERC1155Inve
     }
 
     /// @inheritdoc IERC1155Inventory
-    function collectionOf(uint256 nftId) external pure virtual override returns (uint256) {
-        require(nftId.isNonFungibleToken(), "Inventory: not an NFT");
-        return nftId.getNonFungibleCollection();
+    function collectionOf(uint256 nftId) external view virtual override returns (uint256) {
+        require(nftId.isNonFungibleToken(_collectionMaskLength), "Inventory: not an NFT");
+        return nftId.getNonFungibleCollection(_collectionMaskLength);
     }
 
     /// @inheritdoc IERC1155Inventory
@@ -118,7 +125,7 @@ abstract contract ERC1155InventoryBase is ManagedIdentity, IERC165, IERC1155Inve
 
     /// @inheritdoc IERC1155InventoryTotalSupply
     function totalSupply(uint256 id) external view virtual override returns (uint256) {
-        if (id.isNonFungibleToken()) {
+        if (id.isNonFungibleToken(_collectionMaskLength)) {
             return address(uint160(_owners[id])) == address(0) ? 0 : 1;
         } else {
             return _supplies[id];
@@ -135,14 +142,14 @@ abstract contract ERC1155InventoryBase is ManagedIdentity, IERC165, IERC1155Inve
      * @param collectionId Identifier of the collection.
      */
     function _createCollection(uint256 collectionId) internal virtual {
-        require(!collectionId.isNonFungibleToken(), "Inventory: not a collection");
+        require(!collectionId.isNonFungibleToken(_collectionMaskLength), "Inventory: not a collection");
         require(_creators[collectionId] == address(0), "Inventory: existing collection");
         _creators[collectionId] = _msgSender();
         emit CollectionCreated(collectionId, collectionId.isFungibleToken());
     }
 
     function _creator(uint256 collectionId) internal view virtual returns (address) {
-        require(!collectionId.isNonFungibleToken(), "Inventory: not a collection");
+        require(!collectionId.isNonFungibleToken(_collectionMaskLength), "Inventory: not a collection");
         return _creators[collectionId];
     }