+++ contracts/token/ERC1155/ERC1155InventoryIdentifiersLib.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.7.6 <0.8.0;
 
 /**
  * @title ERC1155InventoryIdentifiersLib, a library to introspect inventory identifiers.
- * @dev With N=32, representing the Non-Fungible Collection mask length, identifiers are represented as follow:
+ * @dev With 0 < N < 256 representing the Non-Fungible Collection mask length, identifiers are represented as follow:
  * (a) a Fungible Token:
  *     - most significant bit == 0
  * (b) a Non-Fungible Collection:
@@ -18,19 +18,19 @@ library ERC1155InventoryIdentifiersLib {
     // Non-Fungible bit. If an id has this bit set, it is a Non-Fungible (either Collection or Token)
     uint256 internal constant _NF_BIT = 1 << 255;
 
-    // Mask for Non-Fungible Collection (including the nf bit)
-    uint256 internal constant _NF_COLLECTION_MASK = uint256(type(uint32).max) << 224;
-    uint256 internal constant _NF_TOKEN_MASK = ~_NF_COLLECTION_MASK;
-
     function isFungibleToken(uint256 id) internal pure returns (bool) {
         return id & _NF_BIT == 0;
     }
 
-    function isNonFungibleToken(uint256 id) internal pure returns (bool) {
-        return id & _NF_BIT != 0 && id & _NF_TOKEN_MASK != 0;
+    function isNonFungibleToken(uint256 id, uint256 collectionMaskLength) internal pure returns (bool) {
+        return id & _NF_BIT != 0 && id & _getNonFungibleTokenMask(collectionMaskLength) != 0;
     }
 
-    function getNonFungibleCollection(uint256 nftId) internal pure returns (uint256) {
-        return nftId & _NF_COLLECTION_MASK;
+    function getNonFungibleCollection(uint256 nftId, uint256 collectionMaskLength) internal pure returns (uint256) {
+        return nftId & ~_getNonFungibleTokenMask(collectionMaskLength);
+    }
+
+    function _getNonFungibleTokenMask(uint256 collectionMaskLength) private pure returns (uint256) {
+        return (1 << (256 - collectionMaskLength)) - 1;
     }
 }