+++ test/contracts/token/ERC1155/ERC1155InventoryPausableMock.test.js
@@ -8,6 +8,7 @@ const implementation = {
   nfMaskLength: 32,
   revertMessages: {
     // ERC1155
+    WrongCollectionMaskLength: 'Inventory: wrong mask length',
     SelfApprovalForAll: 'Inventory: self-approval',
     ZeroAddress: 'Inventory: zero address',
     NonApproved: 'Inventory: non-approved sender',
@@ -65,10 +66,10 @@ const implementation = {
       return contract.createCollection(collectionId, overrides);
     },
   },
-  deploy: async function (deployer) {
+  deploy: async function (deployer, collectionMaskLength = 32) {
     const registry = await artifacts.require('ForwarderRegistry').new({from: deployer});
     const forwarder = await artifacts.require('UniversalForwarder').new({from: deployer});
-    return artifacts.require('ERC1155InventoryPausableMock').new(registry.address, ZeroAddress, {from: deployer});
+    return artifacts.require('ERC1155InventoryPausableMock').new(registry.address, ZeroAddress, collectionMaskLength, {from: deployer});
   },
   mint: async function (contract, to, id, value, overrides) {
     return contract.methods['safeMint(address,uint256,uint256,bytes)'](to, id, value, '0x', overrides);